requirements.txt
flask
dotenv-python
google-genai